Renaming Components folder to components to avoid React conflicts.
